requirements.txt
amqp==5.1.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.5
colorlog==3.1.4
contourpy==1.1.1
cvxopt==1.3.2
cycler==0.11.0
easydict==1.10
ffmpeg-python==0.2.0
filelock==3.12.4
fonttools==4.42.1
future==0.18.3
google-api-core==1.34.0
google-auth==2.23.1
googleapis-common-protos==1.60.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
imutils==0.5.4
install==1.3.5
is-msgs==1.1.18
is-wire==1.2.1
Jinja2==3.1.2
kiwisolver==1.4.5
lit==16.0.6
llvmlite==0.41.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mmdeploy-runtime-gpu==1.3.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.1
numba==0.58.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencensus==0.5.0
opencensus-context==0.1.0
opencv-python==4.8.0.76
opencv-contrib-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.1
prometheus-client==0.3.1
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.184
urllib3==2.0.5
vine==5.0.0
zipp==3.17.0
tensorrt==8.6.1.post1
tensorrt-dispatch==8.6.1.post1
tensorrt-lean==8.6.1.post1